refactor!: remove in-library auth (callers inject http.Client)
mcp-go never handles auth — tokens, OAuth flows, and credentials live
entirely in the caller-supplied http.Client transport (client) or at
the transport/proxy layer (server). Per the non-negotiable, all
in-library auth is removed:

- Delete middleware/auth.go and all its symbols (Auth, Authenticator,
  Identity, APIKey/BearerToken authenticators, StaticAPIKeys,
  StaticTokens, ChainAuthenticators, OAuth2Authenticator, JWTValidator,
  IdentityFromContext, ContextWithIdentity, WithAuth* options).
- Remove the mcp.go auth re-exports and the BearerAuth helper.
- Remove client.WithBearerToken; auth is injected via WithHTTPClient.
- Move the non-auth tracing tests out of middleware/enterprise_test.go
  into middleware/tracing_test.go (the deleted file mixed both).
- Rewrite the list-filter context test and the mTLS doc example to use
  a caller-defined context key instead of the removed Identity type.
- Strip Auth/OAuth2 feature copy from README, website, and docs.

BREAKING CHANGE: in-library authentication is removed. Replace
client.WithBearerToken(tok) with a custom http.RoundTripper on
mcp.WithHTTPClient. Server-side auth must be terminated at the
transport/proxy layer or in caller-provided middleware. See CHANGELOG
for the full migration note.
