+++ website/src/pages/index.astro
@@ -221,7 +221,7 @@ func main() {
         </div>
         <div class="feature">
           <h3>Enterprise Middleware</h3>
-          <p>Audit logging, distributed tracing with correlation IDs, and OAuth2/JWT authentication.</p>
+          <p>Audit logging, distributed tracing with correlation IDs, rate limiting, and size limits.</p>
         </div>
       </div>
